requirements.txt
torch>=2.4
torchvision>=0.19
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
SimpleITK>=2.3
Pillow>=9.5
scipy>=1.10
scikit-image>=0.22
tqdm>=4.66
